Add files via upload
topic is "how to assign a geographic role to a Data field".

By default the first actions done by Tableau when we bring in any data source is to assign a data type and a role (dimension or measure).

But sometimes the data type that is auto-assigned by Tableau may not be correct. In this case we will need to manually correct the data type e.g.: changing from String to geographic. 

Given below are the various methods by which we can assign a geographic role to a data field:

METHOD-I: IN THE PREVIEW DATA SOURCE VIEW

Step 1: Select the column or field for which the data type need to be changed to a geographic role 

Step 2: Click on the data type icon on the Top Left side of the column
     
Step 3: Click on "Geographic Role" 
Click on the required geographic role e.g.: "City"

Step 4: This step is only for confirmation that geographic field was correctly assigned.
    Confirm that the icon on the Top Left side of the column changes to a Globe (signifies a geographic field)

METHOD-II: IN THE MANAGE METADATA VIEW 

Step 1: Select the row or field for which the data type need to be changed to a geographic role 

Step 2: Click on the data type icon on at the Left side of the row
     
Step 3: Click on "Geographic Role" 
Click on the required geographic role e.g.: "City"

Step 4: This step is only for confirmation that geographic field was correctly assigned. 
  Confirm that the icon on the Left side of the row changes to a Globe (signifies a geographic field)

METHOD-III: IN THE DATA PANE

Step 1: Select the field for which the data type need to be changed to a geographic role

Step 2: Right-click the field 

Step 3: Click on "Geographic Role" 
Click on the required geographic role e.g.: "City"

Step 4: This step is only for confirmation that geographic field was correctly assigned. 
    Confirm that the icon on the Left side of the row changes to a Globe (signifies a geographic field)