Add files via upload
This topic is "how to change the default aggregation for a measure".

We will be looking into two methods to change the default aggregation for a measure.
However the steps mentioned here are only for a reference purpose.
There are multiple ways by which we can change the default aggregation for a measure.

Given below are the two methods by which we can change the default aggregation for a measure

METHOD-I: IN THE VIEW AREA (WITH THE MEASURE PILL IN EITHER ROWS OR COLUMNS OF THE VIEW)

Assume that we have a Viz showing the Sales across Categories

Step 1: By default, the aggregation for sales will be SUM(Sales).

Step 2: Right-click the data field (i.e. SUM(Sales))

Step 3: Select the required aggregation e.g.: "Average" 
     
Step 4: Ensure that the updated aggregation is visible in Rows/Columns of the view i.e. AVG(Sales)

NOTE: The aggregation change that is done by this method will be limited only to that particular worksheet/visual.


METHOD-II: IN THE DATA PANE (BEFORE MAKING USE OF THE MEASURE IN EITHER ROWS OR COLUMNS OF THE VIEW)

Step 1: Select the measure for which the aggregation needs to be changed.
    
Step 2: By default, the aggregation for sales will be Sum.

Step 3: Select the required aggregation e.g.: "Average" 

Step 4: If we make use of the Sales measure in a new view either in Rows or Columns the aggregation level will be the updated one i.e. AVG(Sales)

NOTE: The aggregation change that is done by this method will be applicable to all further NEW worksheets/dashboards/stories.