Add files via upload
This topic is "How to change the default color for a data field".

We will be looking into only two methods i.e. directly changing the color and applying the Categorical method of changing the color of a data field.
These steps are meant only as a reference.
There are multiple ways by which we can change the default color for a data field.

Given below are the steps for changing the default color for a data field (in this case a dimension e.g.: Category)

METHOD-I: Directly changing the Color in the Marks card

Assume that we have a Viz showing the Sales across Categories

Step 1: By default, the color for the Categories will be Blue.

Step 2: Click on "Color" of the "Marks" card

Step 3: Select the required color 

Step 4: Ensure that the updated color is displayed for the data field in this case Category

METHOD-II: By dropping a dimension into the Color section of Marks card:

Step 1: By default, the color for the Categories will be Blue.

Step 2: Drop the dimension in this case Category on the Color section of the "Marks" card

Step 3: To change the automatic colors assigned click on "Color" of "Marks"
    Click on "Edit Colors..." 

Step 4: In the "Select Data Item" click on the item for the color is to be changed.

Step 5: Choose the required color from the Right-Side Color Palette. 
    Click "OK"

Step 6: Ensure that the updated color is displayed for the member in data field in this case Furniture Category