Add files via upload
This topic is "How to change the default date format".

We will be looking into only one method i.e. changing the date format for a date dimension.
These steps are meant only as a reference.
There are other ways by which we can change the date format.

Given below are the steps for changing the Date Format for a Date dimension (in this case it is Order Date)

Assume that we have a Viz showing the trend of Profit over Order Date (Continuous Date Value set to Month)

Step 1: By default, the Date Format will be set to "Automatic" e.g. Month of Order Date: November 2017
    This can be confirmed by hovering over any one of marks in the Line chart.

Step 2: Select the dimension in the Data pane in this case it is Order Date

Step 3: Right-click on Order Date

Step 4: Select Default Properties and Date Format...

Step 5: Confirm that the Default Date Format is "Automatic"

Step 6: Select the required Date Format e.g.: "Wednesday, 14 March, 2001"

Step 7: Ensure that the updated Date Format is being displayed.
    This can be confirmed by hovering over the same mark that was checked in Step 1.