Add files via upload
This topic is "How to create a Line Chart".

The Line Chart falls in the Trend Visual Segment.
A Line Chart is typically used to show the trends of a data field over time E.g.: Sales between certain years 2018 to 2021 as per a Date Field.

In case we have want to increase the granularity of the line chart we can drop a dimension e.g.: Segment to the Color section of Marks card. The view will change based on the number of members in the dimension that was added. Hence the number of marks in the viz will increase.

By default when any date is dropped into the Rows/Columns it will be in Discrete Date Part format in Years (Blue Color) e.g.: YEAR(Order Date)
In this case the Date field will form a Header since it is discrete.

If required we can convert it to Continuous Date Value (Green Color) e.g.: MONTH(Order Date)
In this case the Date field will form an Axis.

We will go through the typical steps involved in creating a Line Chart without using the "Show Me" option.