Add files via upload
This topic is "How to create a Scatter Plot".

The Scatter Plot falls in the Relation Visual Segment.
A Scatter Plot is typically used to show the relationship between measures E.g.: Relationship between Sales and Profit.
Hence Scatter Plots are used for Correlation Analysis to identify the relationship between measures.

It is a general practice of plotting the independent variable or measure in the X-Axis (Columns Shelf) and the dependent variable or measure in the Y-Axis (Rows Shelf).

By default the measures will be aggregated to SUM.
In case we have want to increase the granularity of the scatter plot we can drop a dimension e.g.: Product Name to the Details section of Marks card.
The view will change based on the number of members in the dimension that was added. Hence the number of marks in the viz will increase.

In Scatter Plots we can make use of 2 or more attributes e.g.: Size and Color for enabling better correlation analysis.

Given below are the basic steps to create a Scatter Plot without using the "Show Me" option:

Assume that we need to find the relationship between Sales and Profit

Step 1: Drag and drop the 1st measure (independent field) e.g.: Sales in the Columns shelf

Step 2: Drag and drop the 2nd measure (dependent field) e.g.: Profit in the Rows shelf

Step 3: By default, both measures will be aggregated. Hence there will be only 1 Mark in the Viz. 
This can be confirmed by checking the mark count in the bottom left side of the view pane.

Step 4: Drag and drop any dimension e.g.: Product Name to the Details section of Marks card
    This will result in a more granular view that can be confirmed by checking the marks count in the bottom left side of the view pane.
Profit Values having negative values will be enclosed in brackets.

Step 5: In case we need to add Size attribute to the Scatter Plot drag and drop a measure e.g.: Sales to the Size section of Marks card

Step 6: In case we need to add Color attribute to the Scatter Plot drag and drop a dimension e.g.: Region to the Colors section of Marks card

Step 7: In case we want to analyze only one region we can make use of the "Show Highlighter" option
    Right-click on the dimension e.g.: Region 

Step 8: Click on the "Show Highlighter" 

Step 9: Confirm that the Highlight box is visible in the Right Side of the view pane

Step 10: From the drop-down of the Highlighter box select the required member of dimension e.g.: Central

Steps 5 to 10 are optional and maybe used for better correlation analysis.