Add files via upload
This topic is "How to create a Filled Map using Geographic data".

There are many types of Maps that can be created in Tableau but we are going to focus only on the two main types
a) Symbol Maps
b) Filled Maps (also called Choropleth Maps)

In this post we will go through the steps of creating a Filled Map.

The pre-requisite to creating maps in Tableau is that the data source should contain geographic data i.e. the data type of the field should be indicated by the globe symbol.

Given below are the basic steps to create a Filled Map without using the "Show Me" option:

Step 1: Drag and drop any Geographic field e.g.: State into the View area
In case you are located in an area that is different from the country/area of the data source you will see a message like "49 unknown" at the bottom right side corner of view area
Click on the unknown message

Step 2: A new window called like "Special Values for [State]" will open
Click on "Edit Locations..."

Step 3: A new window called like "Edit Locations" will open

Step 4: Click on the "Country/Region" field (near drop-down)

Step 5: Click on the "Fixed" field (near drop-down)

Step 6: Enter the correct country name e.g.: United States
Click on OK

Step 7: Now we must see the actual map of country in this case United States with the a circle representing a central point of each State
Whenever we bring a geographic field into the view area, Tableau compares the field with respect to it's internal geo-coded database.
Based on the comparison a Latitude and Longitude are assigned on the map and the geographic field is assigned to the Detail section of the Marks card

Step 8: Drag and drop any measure e.g.: Sales to the Color section of the Marks card
A color legend will be present in the Top Right side of View Area.
The darker color represents the States with Higher Sales while the lighter color represents States with Lower Sales.

Step 9: Drag and drop the Geographic field in this case State to the Label section of the Marks card

Step 10: Drag and drop the measure in this case Sales to the Label section of the Marks card

Step 11: The State with darkest color has the highest sales in this case it is California with a Sales of $458K

Step 12: The State with lightest color has the lowest sales in this case it is Maine with a Sales of $1K

Step 13: If required we can change the Background of Map by selecting the required option using the Map --> Background Maps

Step 14: If required we can change the Map Layers by first selecting the required option using the Map --> Map Layers
In the "Map Layers" pane we can select or de-select the required options e.g.: Terrain, Coastline etc.