Add files via upload
This topic is "How to create a Symbol Map using Geographic data".

In the previous post we had gone through the Filled Maps.

In this post we will go through the steps of creating a Symbol Map.

NOTE: The topic of creating multi-layered maps will be covered in a separate post/article.

The pre-requisite to creating maps is that the data source should contain geographic data i.e. the data type of the field should be indicated by the globe symbol

Given below are the differences between Symbol and Filled Maps.
The symbol maps use the symbols to represent a central point of a geographic region whereas in filled map boundaries of a geographic region is filled with a color.
With a Filled map we can typically encode only one measure as Color e.g.: Sales 
But in a Symbol map we can typically encode at least 2 measures i.e. 1 in Size marks card and 1 in Color marks card e.g.: Sales and Profit.
So in a Symbol Map the Size can represent the Sales and Color of the symbol represents the Profit.