Add files via upload
This topic is "How to Format view for presentation "

Formatting is an important part of both the analysis and presentation. 
We can format almost everything on a worksheet including the fonts, shading, alignment, borders, and graph lines. 
For example, in a text table we may want to add banded shading to help in visually separating consecutive groups of rows or columns. 
In a scatter view with reference lines you may want to change the line thickness and color. All of these settings can be changed using the Format window.

We will be going through the below mentioned sub-topics under Format view:
1. Use color from the marks card
2. Configure fonts
3. Format marks as shapes
4. Change size of marks
5. Configure viz animations
6. Show and hide legends

The examples that I will be covering will be limited to only few options for each sub-topic.
For more detailed information please go through the tableau online help links.

The exact steps are mentioned in the attached document.