Add files via upload
This topic is "How to create and use a Parameter"

A parameter is a workbook variable such as a number, date, or string that can replace a constant value in a calculation, filter, or reference line.
Parameters are used to provide more control to the users. For example we can use parameters to display Top N products where N can be configurable via a parameter.

Important steps when using Parameters:

Step 1: Create the Parameter
Step 2: Use the Parameter
Step 3: Create calculated field (optional)
Step 4: Show the Parameter (optional)

There are multiple methods to create parameters.
We will go through one particular manner to create a Top N parameter from the filter option of a dimension.

The parameters can be made more dynamic and interactive by using them in Parameter Actions. 
Parameter actions let the end user change a parameter value through direct interaction with a viz, such as clicking or selecting a mark.

The exact steps are mentioned in the attached document.