Add files via upload
This topic is "How to create a calculated field (e.g. string, date, simple arithmetic)"

On some occasions the data source/data set will not contain a field (or column) that might be needed for analysis. 
For example, the data source might contain fields with values for Sales and Profit, but not for Profit Ratio or Profit Margin. 
If this is the case, we can create a calculated field for Profit Ratio using data from the Sales and Profit fields.

We will go through examples of how to create simple calculated fields for string, date and simple arithmetic options.

The exact steps are mentioned in the attached document.