Add files via upload
This topic is "How to use a Quick Table calculation".

Quick table calculations allow us to quickly apply a common table calculation to the visualization using the most typical settings for that calculation type. 
They are special calculated fields that are computed only on the measures in the current view/ visualization.

The following quick table calculations are available in Tableau:
Running total
Difference
Percent difference
Percent of total
Rank
Percentile
Moving average
YTD total
Compound growth rate
Year of year growth
YTD growth

When we add a table calculation, we must use all dimensions in the level of detail either for partitioning (scoping) or for addressing (direction).
The dimensions that define how to group the calculation (the scope of data it is performed on) are called partitioning fields. 
The table calculation is performed separately within each partition.
The remaining dimensions, upon which the table calculation is performed, are called addressing fields, and determine the direction of the calculation.

Some examples of Scope and direction in Table Calculation are Table (across), Table (down), Pane (across), Pane (down), Cell etc.

Quick table calculations are table calculations that we can apply quickly to the visualization in Tableau. 
They are applied to the visualization with the most typical settings for the calculation type we choose so that we can continue on with the analysis. 
With traditional table calculations, we can apply the same settings, but we must apply them manually.

Table Calculations can be used in any type of visualization but for our example we will be using it in a simple crosstabs or text table.
In this post we will be going through only one example on how to apply a quick table calculation to a visualization.

The exact steps are mentioned in the attached document.