Add files via upload
This topic is "How to add a Trend Line to a view".

Trend lines are used in a visualization to highlight trends in the data. 
For example in a Scatter Plot the trend line can be used to determine the relationship between 2 measures e.g.: Sales and Profit

When we add trend lines to a view, we can specify how we want them to look and behave.

These model types are available for trend lines: Linear, Logarithmic, Exponential, Power, and Polynomial.

In this post we will only be going through an example for adding Trend Line to a Scatter Plot.

The exact steps are mentioned in the attached document.