Add files via upload
This topic is "How to Add Reference Line to a view".

We can add reference line, band, distribution, or box plot to identify a specific value, region, or range on a continuous axis e.g.: Measure in a Tableau view. 
For example, if we are analyzing the monthly sales for several products, we can include a reference line at the average sales mark.
This will give us idea as how each product performed against the average sales.

Tableau lets us add as many reference lines, bands, distributions, and box plots to a view as we require.
Reference lines are computed only based on the fields that are present in the current view.

Given below are the different options when using references in Tableau:

Reference Lines - A reference line can be added at a constant or computed value on the axis. 
Computed values can be based on a specified field e.g.: Sales. 
We can also include confidence intervals with a reference line.

Reference Bands - Reference bands shade an area behind the marks in the view between two constant or computed values on the axis.

Reference Distributions - Reference distributions add a gradient of shading to indicate the distribution of values along the axis. 
Distribution can be defined by percentages, percentiles, quantiles , or standard deviation.
Reference distributions can also be used to create bullet charts.

Box Plots - Box plots (also known as box and whisker charts) are a standardized graphic for describing the distribution of values along an axis. 
Box plots show quartiles (also known as hinges) and whiskers. 
Tableau provides different box plot styles, and allows us to configure the location of the whiskers and other details.

In this plot we will only be going through an example for adding Reference Lines to a view.