Add files via upload
This topic is "How to Add a Manual or Computed Sort".

There are many ways we can sort data in Tableau.
When viewing a visualization, data can be sorted using single click options from an axis, header, or field label. 
In the authoring environment, additional sorting options include sorting manually in headers and legends, using the toolbar sort icons, or sorting from the sort menu.

In this post we will be covering Computed and Manual Sorts.

The exact steps are mentioned in the attached document.