Add files via upload
This topic is "How to Add a Date Filter".

When a date field is dragged from the Data pane to the Filters shelf in Tableau, we get various options to choose from in Filter Field dialog box 

We can select whether we want to filter on a relative date; filter between a range of dates; or select discrete dates or individual dates to filter from the view.

Filter relative dates: Click Relative dates to define a range of dates that updates based on the date and time you open the view. 
For example, we may want to see Year to Date sales, all records from the past 30 days, or bugs reported/closed last week. 
Relative date filters can also be relative to a specific anchor date rather than today (default option).

Filter a range of dates: Select Range of dates to define a fixed range of dates to filter. 
For example, we may want to see all orders placed or quantities sold between January 1, 2018 and August 26, 2020.

Filter discrete dates: Select a discrete date value in the dialog box if we want to include entire date levels. 
For example, if we select specific Months, we can choose to filter specific Months (e.g. April, May, June) from the view, regardless of the year.

Latest date preset: If we want to ensure that only the most recent date in a data source is selected in the filter when the workbook is shared or opened, 
select a discrete date such as Month/Day/Year or Individual Dates and then, on the General tab, select Filter to latest date value when workbook is opened.

Filter individual dates: Select Individual dates to filter specific dates from the view.

Additional date filter options: When we select Relative dates or Range of dates, the Filter dialog box opens. 
In that dialog box, we can define a Starting date or Ending date. We can also select Special to include null dates, non-null dates, or all dates.

The exact steps are mentioned in the attached document.