Add files via upload
This topic is "How to Add a Context Filter".

By default, all filters that we set and use in Tableau are computed independently. That is, each filter accesses all rows in the data source without regard to other filters. 

However, we can set one or more categorical filters as context filters for the view. The context filter can be considered as being an independent filter. Any other filters that we set are defined as dependent filters because they process only the data that passes through the context filter.

Typical uses of context filters are mentioned below:

Improve performance – In case we have a lot of filters or have a large data source, the queries can be slow. 
We can set one or more context filters to improve performance.

Create a dependent numerical or top N filter – We can set a context filter to include only the data of interest, and then set a numerical or a top N filter.
We will be going through a top N filter after creating a context filter

Context Filters can be used in all types of visuals even maps.
We will making use of a Context Filter in a standard Crosstabs view.
Consider that we want to identify the Top 10 Cities in California State by Sales. For this case we need to make use of a Context Filter.

The exact steps are mentioned in the attached document.