Add files via upload
This topic is "How to Add Basic Filters to Views".

Filtering plays an important part in analyzing data. 
We will go through the many ways that we can filter data from the view. 
We can display interactive filters in the view, and format filters in the view.

As a prerequisite to filtering data in Tableau, it's important to understand the order in which Tableau executes filters in the workbook.

Tableau performs actions on the view in a very specific order; this is called the Order of Operations. 
Filters are executed in the following order:

1. Extract filters
2. Data source filters
3. Context filters
4. Filters on dimensions (whether on the Filters shelf or in filter cards in the view)
5. Filters on measures (whether on the Filters shelf or in filter cards in the view)

In this post we will go through the Dimension and Measure Filters alone.
Data source and Context filters will be covered in separate posts.
Extract filters are currently out of scope of these posts since it is available only in Tableau Desktop Professional Edition.

The exact steps are mentioned in the attached document.