Add files via upload
This topic is "How to create Hierarchies".

When we connect to a data source, Tableau automatically separates date fields into hierarchies so that we can easily break down the viz. 
Date Hierarchies will be explained in a separate post.
We can also create our own custom hierarchies. 
For example, if we have a set of fields named Region, State, and County, we can create a hierarchy from these fields so that we can quickly drill down between levels in the viz.

Given below are the main methods to create a Hierarchies:
Method I: Right-click the Dimension from the Data Pane and click Hierarchy --> Select Hierarchy 
Method II: By dropping the Dimension on to another Dimension in the Data Pane

The exact steps are mentioned in the attached document.