Add files via upload
This topic is "How to create Sets".

Sets are useful to view and highlight the data that meet a particular condition or criteria.

Sets can be used to compare and ask questions about a subset of data. 
Sets are custom fields that define a subset of data based on some conditions.

Sets and Groups are very similar but there are some key differences between them.

1. The most significant difference is that sets are dynamic in nature while groups are not. 
This means that when the underlying data changes the dynamically created sets will update with it while this is not an option with groups.

2. Sets offer greater flexibility as we can link them to a condition. 

3. Groups can be formed only within one dimension while with sets can be formed across multiple dimensions

Sets can be formed by using the Venn Diagram icon ((two interlocking rings) inside the Tooltip menu that appears when you point at any of the selected marks.

There are two types of sets: dynamic sets and fixed sets. 
The members of a dynamic set change when the underlying data changes. 
Dynamic sets can only be based on a single dimension.
The members of a fixed set do not change, even if the underlying data changes. 
A fixed set can be based on a single dimension or multiple dimensions.

Given below are the main methods to create a Set:
Method I: From the Data Pane (Dynamic Sets)
Method II: By using Marks (directly from the view/ visualization) (Fixed Sets)

The exact steps are mentioned in the attached document.