Add files via upload
This topic is "How to create Groups".

A Tableau Group is a collection of multiple related members combined in a single dimension to create a higher category of the dimension. 
Tableau allows the grouping of single-dimensional members and automatically creates a new dimension by adding group at the end of the name (within round brackets). 
Tableau does not do anything with the original dimension of the members.

Groups are used to combine several members of a dimension into a new dimension that did not originally exist in the data source.
Groups can be created from the Data Pane or directly from the view.
Typically if the data points or members are more it is preferred to create the Groups from the Data Pane.
However if we have lesser data we can create groups directly from the visualization.

Marks can be grouped by using the "paper clip" icon inside the Tooltip menu that appears when you point at any of the selected marks.

Given below are the methods to create a Group:
Method I: From the Data Pane 
Method II: By using Marks (directly from the view/ visualization)
Method III: By using Headers (directly from the view/ visualization)

The exact steps for all 3 methods are mentioned in the attached document.