Add files via upload
This topic is "How to create a Crosstabs".

Crosstabs are also known as "Text tables".
They are used to display or represent data in the view similar to that of a text file or excel spreadsheet.
Typically Crosstabs are used for a quick verification of data.
Crosstabs are less commonly used in Visuals and Dashboards as they do not provide a pictorial representation of the data.

Given below are the basic steps to create a Crosstabs or Text Table without using the "Show Me" option:

Step 1: Drag and drop any dimension e.g.: Ship Mode to the Rows or Columns Shelf

Step 2: Drag and drop the 2nd dimension e.g.: Category to the Rows or Columns Shelf
    
Step 3: Drag and drop any measure e.g.: Sales to the Text section of the Marks card or to the View

Step 4: Drag and drop the 2nd measure e.g.: Profit to the View 
     
Step 5: Ensure that both measures are present in crosstabs view.
Once the 2nd measure is moved into the view Tableau will automatically place Measure Names in Columns and Filter Shelf, Measure Values in Text Label.
A separate shelf called Measure Values will also appear.

Step 6: Right-Click on the Measure Names in the Filter Shelf and select "Edit Filter..."

Step 7: Ensure that a new window "Filter [Measure Names]" is displayed

Step 8: Select the additional Measure Names e.g.: Discount and Quantity and select OK

Step 9: Ensure that the crosstabs now has the newly included measure names. 
It may be required to increase the size of the table to ensure that all column names are displayed fully.

Step 10: Right-click on the Crosstabs and select “Format”.

Step 11: Under Format Borders click on Borders --> Rows

Step 12: Change the color of the Row Divider to the required color

Step 13: Ensure that the crosstabs now has the updated Row Divider color.

Step 14: Similarly change the color of the Column Divider to the required color

Step 15: Ensure that the crosstabs now has the updated Column Divider color.

Step 16: If it is required, we can add the column grand totals using
Analysis --> Totals --> Show Column Grand Totals

Step 17: Ensure that the Grand Total is now visible in the crosstabs

Step 18: If it is required, we can add all the Subtotals using
Analysis --> Totals --> Add All Subtotals

Step 19: Ensure that the Sub Totals are now visible in the crosstabs

Step 20: If required further formatting can be done by using the relevant options under Format pane and are not limited to the mentioned options shown in the example.

Steps 10 to 20 are meant only for additional formatting of crosstabs and are not limited to the mentioned options shown in the example.

The exact steps are mentioned in the attached document.