Add files via upload
This topic is "How to create a Highlight Table".

Highlight Tables are a variation of the Crosstabs.
Highlight Tables makes use of a measure on Text and Color section of Marks card.

When we drop a field with continuous values on the Marks card (typically a measure), Tableau displays a quantitative legend with a continuous range of colors by default light blue to dark blue.

The advantage of the Highlight Table is that it does not take away the spreadsheet or text table that many or most of us are already familiar with.
In addition to this the Highlight Table improves the value of the visualization by using color to make the biggest and smallest values really stand out. 
We can also identify the Row item e.g.: Sub-Category is doing the best and worst w.r.t the Column item e.g.: Segment.
Highlight Tables are one of the ways to help people GET data visualization as it opens the doors to better data visualization.
Most of these points are taken from Steve Wexler's LinkedIn course "How to use Data Visualization to Make Better Decisions -Faster" Section - Color plus numbers = amazing data visualization.

Given below are the basic steps to create a Highlight Table without using the "Show Me" option:

Step 1: Drag and drop any dimension e.g.: Sub-Category to the Rows or Columns Shelf

Step 2: Drag and drop the 2nd dimension e.g.: Segment to the Columns or Rows Shelf
    
Step 3: Drag and drop any measure e.g.: Profit to the Text section of the Marks card or to the View

Step 4: Drag and drop the same measure e.g.: Profit to the Color section of the Marks card
Ensure that the Color Legend is present in the Top Right corner of View Area

Step 5: Select the drop-down as "Square" below the Marks 

Step 6: It may be required to increase the size of the table to ensure that all column names are displayed fully.
Quick takeaways are the Tables have the lowest Profit with ($10K) and Phones have the highest Profit with $24K
Copiers have high profits in all Segments.

The exact steps are mentioned in the attached document.