Add files via upload
This topic is "How to create a Stacked Bar Chart".

Like the Bar Chart the Stacked Bar Chart also falls in the Comparison Visual Segment.

The stacked bar chart is the type of the bar chart with segmented bars. 
For E.g.: Each bar can represent the different regions for each sub-category.

The exact steps are mentioned in the attached document.