Add files via upload
This topic is "How to create a Dual Axis Chart".

Like the Bar Chart & Combined Axis Chart, the Dual Axis Chart also falls in the Comparison Visual Segment.

As the name suggests the chart contains two separate and independent axes. 
This chart allows us to compare multiple measures over a common field e.g.: Order Date. 
First we need to make use of a dimension(Label/Header) e.g.: Order Date in the Columns or Rows Shelf 
Next we need to make use of two different measures e.g.: Profit and Sales in the Rows or Columns Shelf. 
In case both measures are having the same unit of measurement e.g.: Sales and Profit we need to synchronize both the axes.
We can have different marks types for each measure e.g.: Sales can be changed to Bar marks type and Profit can be changed to Line marks type.
In this case the chart is known as "Dual Axis Combination Chart" as multiple mark types are used in the same visualization.

The exact steps are mentioned in the attached document.