Add files via upload
This topic is "How to create a Combined Axis Chart".

Like the Bar Chart, the Combined Axis Chart also falls in the Comparison Visual Segment.

As the name suggests the chart combines the Axis for 2 or more measures. It is created with one shared axis.
First we need to make use of a dimension(Label/Header) e.g.: Region in the Columns or Rows Shelf 
Next we need to make use of two different measures e.g.: Profit and Sales in the Rows or Columns Shelf. 
Placing measures on the same axis allows us to build a view in which we can easily compare those measures using a single chart/viz.

The exact steps are mentioned in the attached document.