Add files via upload
This topic is "How to create a Box Plot".

Box plots, also known as box-and-whisker plots, are used to show and compare the distribution of values along an axis.
They are also useful in identifying outliers.

The Boxes of the Box Plot indicate the middle 50 percent of the data (that is, the middle two quartiles of the data's distribution).
The remaining 25% of the data is below the box and other 25% of the data is above the box.
The width or height of the box is called the "interquartile range", abbreviated "IQR". That is, IQR = Upper Hinge – Lower Hinge.
The IQR can be used as a measure of how spread-out the values are. Statistics assumes that your values are clustered around some central value.
An outlier is any value that lies more than one and a half times the length of the box from either end of the box.

The lines in the box plot also called whiskers are used to display all points within 1.5 times (configurable) the interquartile range (in other words, 
all points within 1.5 times the width of the adjoining box), or all points at the maximum extent of the data.

The upper whisker of the box plot is the largest dataset number smaller than 1.5IQR above the third quartile (or Upper Hinge)
Similarly, the lower whisker of the box plot is the smallest dataset number larger than 1.5IQR below the first quartile (or Lower Hinge)

Outliers are any data points that lie above Upper Hinge + 1.5 X (IQR) or below Lower Hinge - 1.5 X (IQR)

The basic building blocks for a box and whisker plot are as follows:

Mark type: Circle
Columns shelf: Dimension
Rows shelf: Measure
Detail: Dimension
Reference Line: Box Plot

The attached document provides the detailed steps to create a Box Plot without using the "Show Me" option