Add files via upload
This topic is "How to create a Bar in Bar Chart".

Bar in bar charts are used to compare the measures against the goals, similar to Bullet graphs. e.g.: Profit Vs Target Profit
This chart is actually built on top of a combined axis chart.

Bar In Bar Charts like Bar Charts fall under the Comparison Visual Segment.

Given below are the basic steps to create a Bar In Bar Chart without using the "Show Me" option:

Step 1: Drag and drop the 1st Measure e.g.: Profit to the Columns Shelf

Step 2: Drag and drop the Dimension e.g.: Sub-Category to the Rows Shelf

Step 3: Drag and drop the 2nd Measure e.g.: Target Profit to Profit Axis
NOTE: Target Profit is a Calculated Field with value [Sales]/10
This is the Combined Axis Chart.

Step 4: Drag and drop the "Measure Names" from Rows Shelf to the Color on Marks card
This is to ensure that there is only 1 bar for each Sub-Category  
Please note that a Color Legend appears in the Top Right side of the view

Step 5: Go to Analysis --> Stack Marks --> Off
This is to ensure that both bars start from zero baseline by unstacking the marks

Step 6: Press Ctrl and Drag and drop the "Measure Names" to the Size on Marks card
We can show both bars by changing the size of bars 
Please note that a Size Legend appears in the Top Right side of the view

Step 7: Select the Entire View from the drop-down

Step 8: Now the Bar In Bar Chart view is complete
For this example, we see that Tables, Supplies, Bookcases and Machines are not meeting their Target Profits as for these sub-categories we can see the Profit (Blue) is within or less than the Target Profit (Orange) bar.
  
The exact steps are mentioned in the attached document.