Add files via upload
This topic is "How to create a Bullet Graph".

A bullet graph is a variation of a bar graph developed to replace dashboard gauges and meters. 
A bullet graph is useful for comparing the performance of a primary measure to one or more other measures. 

Bullet Graphs like Bar Charts fall under the Comparison Visual Segment.

Given below are the basic steps to create a Bullet Graph without using the "Show Me" option:

Step 1: Drag and drop the 1st Measure e.g.: Profit to the Columns Shelf

Step 2: Drag and drop the 2nd Measure e.g.: Target Profit to the Detail on the Marks card
NOTE: Target Profit is a Calculated Field with value [Sales]/10

Step 3: Drag and drop the Dimension e.g.: Region to the Rows Shelf

Step 4: Right-click on Profit Axis

Step 5: Click on "Add a Reference Line"

Step 6: Ensure that a new window "Add Reference Line, Band, or Box" is seen

Step 7: Select the Scope as "Cell"

Step 8: Select the Value as "SUM(Target Profit)" from the drop-down

Step 9: Select the aggregation as "Sum" from the drop-down

Step 10: Select the Label as "None" from the drop-down

Step 11: Select the required Line properties

Step 12: Click on "OK"

Step 13: Right-click on Profit Axis and Click on "Add a Reference Line"

Step 14: Select the "Distribution" option

Step 15: Select the Scope as "Cell"

Step 16: Select the Label as "None" from the drop-down

Step 17: Select the "Fill Below" option

Step 18: Click on "OK"

Step 19: Change the Size of Bars using Slider Bar. 
This is to make the shading more visible
 
Step 20: Now the Bullet Graph view is complete
For each region we can decode the below mentioned information
60% of Target Profit is shaded in dark grey
80% of Target Profit is shaded in light grey
100% of Target Profit is shown by black reference line
Actual Profit is represented by the blue bar
For this example we see that Central Region alone is not meeting it's Target Profit.
 
The exact steps are mentioned in the attached document.